requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2
django-environ==0.10.0
djangorestframework==3.14.0
drf-yasg==1.21.5
exceptiongroup==1.1.1
filelock==3.11.0
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
selenium==4.9.1
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
undetected-chromedriver==3.5.3
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.21.0
webdriver-manager==4.0.1
websockets==11.0.3
wsproto==1.2.0
zope.interface==6.0
